Fix grouped query support
For grouped query available in Mistral, it is requires a different H
(conventionally H / Hk) for key / value. Conventionally, H_k is also
smaller than H and H % H_k == 0.

We can leverage these for grouped query implementation. But for now, I
am sticking with the original design to use head_offsets for this. But
we still need to pass the H_k to compute the right offset in case
batched == true.